+++ CHANGELOG.md
@@ -1,1 +1,12 @@
 # Change log
+
+## [v1.0.0-alpha] - [2020-09-01]
+### Added
+- Displaying current profile
+- Displaying list of profiles
+- Adding profiles
+- Switching profiles
+- Updating profiles
+- Removing profiles
+
+[v1.0.0-alpha]: https://github.com/XDragonSl/git-switcher/releases/tag/v1.0.0-alpha